remove all indexing code from snoop
The elasticsearch indexing will be done from search.
Will have to rethink the "resetindex" management command.
